This shit is almost perfect.
Galaxy coordinates are good to go with precision to fucking 0,1°. LST is much more accurate with precision of 0,1 second or even better. It's fucking craftsmanship EXCEPT fucking LST. This fucking shit decide to fuck up when UTC is around 0h. For some reason at midnight when you want to see what LST is the program jump and say it's fucking year 0